backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

#Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

#Computer Vision
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0
ultralytics==8.0.196
torch==2.1.0
torchvision==0.16.0
onnx==1.15.0
onnxruntime==1.16.3

#Utilities
pydantic==2.5.0
aiofiles==23.2.1
httpx==0.25.2
redis==5.0.1
requests==2.31.0

#Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0